// backend/src/core/queues/mail-queue/models/get-details.model.ts
export class GetDetailsOfReservation{
    entryCity: string;
    devolutionCity: string;
    entryDateTime: string;
    devolutionDateTime: string;
    daysOfRent: number;
    createdAt: string;
    vehicle: string;
    transmission: string;
    model: number;
    totalValue: string;
    percentage_of_total_value: string;
    percentage_in_values: string;
    outstandingBalance: string;
    providerName: string;
    providerEmail: string;
    providerPhone: string;
    providerCity: string;
    providerAddress: string;
    customerName: string;
    customerEmail: string;
    customerPhone: string;
    customerIdentification: string;
    customerCity: string;
    customerAddress: string;
}